Reverted a change to EmptyDisallow test
Reverting a change that was made just after I committed this test.

The robots.txt file test isn't multiple tests, its a single one, and with issue #22 now fixed, it fails for no reason. Test yourself in Googles robots.txt tester, it will show you that this test is just fucked up like it is. Reverting to fix it, not to create another!